pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"



[project]
name = "bot-launcher"
version = "0.1.0"
description = "Master bot to launch other bot containers"
authors = [
    {name = "harikrishna2005@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "mqtt-connector-lib @ git+https://github.com/harikrishna2005/mqtt_connector_lib.git@redis_connector_develop",
    "docker (>=7.1.0,<8.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.5,<0.129.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "psutil (>=5.9.0,<7.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]




[tool.poetry.scripts]
run-manager = "bot_launcher.cli:start_launcher"
#run-rebalancing = "bots.rebalancing_bot.main:main"
run-rebalancing = "bots.rebalancing_bot.cli:start_server"
#       = "<package>.<module>:<function>"


[tool.poetry]
package-mode = true
packages = [
    { include = "bot_launcher", from = "src" },
    { include = "bots", from = "src" },
    { include = "connectors", from = "src" },
    { include = "shared", from = "src" }

]



#
## In Poetry 2.0, optional dependencies for extras go here
#[tool.poetry.dependencies]
#pytest = { version = "^7.4.0", optional = true }
#pytest-asyncio = { version = "^0.21.0", optional = true }
#pytest-cov = { version = "^4.1.0", optional = true }
#
#
#[tool.poetry.extras]
#test = ["pytest", "pytest-asyncio", "pytest-cov"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=0.21.0,<0.22.0)",
    "pytest-cov (>=4.1.0,<5.0.0)"
]
